requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_google_genai
langchain-openai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
pandas
streamlit
pytest
langchain-core[tracing]
docx2txt
-e .
